remove OperationQueue
